+++ src/containers/AudioUploader/components/AudioContent.tsx
@@ -17,8 +17,6 @@ import { DeleteForever } from '@ndla/icons/editor';
 import IconButton from '../../../components/IconButton';
 import AudioPlayer from './AudioPlayer';
 import FormikField from '../../../components/FormikField';
-import PlainTextEditor from '../../../components/SlateEditor/PlainTextEditor';
-import textTransformPlugin from '../../../components/SlateEditor/plugins/textTransform';
 import { AudioFormikType } from './AudioForm';
 import { TitleField } from '../../FormikForm';
 
@@ -58,8 +56,6 @@ const getPlayerObject = (
   }
   return undefined;
 };
-
-const plugins = [textTransformPlugin()];
 
 const AudioContent = ({ t, formik }: Props & tType) => {
   const { values, setFieldValue, submitForm, handleBlur } = formik;
@@ -114,26 +110,6 @@ const AudioContent = ({ t, formik }: Props & tType) => {
             </UploadDropZone>
           )
         }
-      </FormikField>
-
-      <FormikField label={t('form.audio.manuscript')} name="manuscript" maxLength={1000}>
-        {({ field }) => (
-          <PlainTextEditor
-            id={field.name}
-            {...field}
-            className={'manuscript'}
-            placeholder={t('form.audio.manuscript')}
-            handleSubmit={submitForm}
-            plugins={plugins}
-            onBlur={(event: Event, editor: unknown, next: Function) => {
-              next();
-              // this is a hack since formik onBlur-handler interferes with slates
-              // related to: https://github.com/ianstormtaylor/slate/issues/2434
-              // formik handleBlur needs to be called for validation to work (and touched to be set)
-              setTimeout(() => handleBlur({ target: { name: 'manuscript' } }), 0);
-            }}
-          />
-        )}
       </FormikField>
     </Fragment>
   );
